@@ -2,4 +2,4 @@
 
 package assert
 
-func Assert(cond bool, msg string, args ...interface{}) {}
+func Assert(cond bool, msg string, args ...any) {}
