@@ -2,7 +2,7 @@ package comparators
 
 // InterfaceEqual говорит, равны ли два интерфейса lhs и rhs.
 // Функция является panic safe.
-func InterfaceEqual(lhs, rhs interface{}) bool {
+func InterfaceEqual(lhs, rhs any) bool {
 	// Реализуй меня.
 	return false
 }
