@@ -9,43 +9,51 @@ import (
 )
 
 func TestMust(t *testing.T) {
-	t.Run("ok nil result nil error", func(t *testing.T) {
+	t.Run("ok no error 1", func(t *testing.T) {
 		assert.NotPanics(t, func() {
-			result := Must(func() (interface{}, error) {
-				return nil, nil //nolint:nilnil
+			result := Must(func() (int, error) {
+				return 42, nil
 			})
-			assert.Nil(t, result)
+			assert.Equal(t, 42, result)
 		})
 	})
 
-	t.Run("ok notnil result nil error", func(t *testing.T) {
+	t.Run("ok no error 2", func(t *testing.T) {
 		cmd := new(exec.Cmd)
 
 		assert.NotPanics(t, func() {
-			result := Must(func() (interface{}, error) {
+			result := Must(func() (*exec.Cmd, error) {
 				return cmd, nil
 			})
 			assert.Equal(t, cmd, result)
 		})
 	})
 
-	t.Run("panic nil result notnil error", func(t *testing.T) {
+	t.Run("ok no error 3", func(t *testing.T) {
+		assert.NotPanics(t, func() {
+			result := Must(func() (*exec.Cmd, error) {
+				return nil, nil //nolint:nilnil
+			})
+			assert.Nil(t, result)
+		})
+	})
+
+	t.Run("panic because error 1", func(t *testing.T) {
 		errExpected := errors.New("something went wrong")
 
 		assert.PanicsWithValue(t, errExpected, func() {
-			Must(func() (interface{}, error) {
-				return nil, errExpected
+			Must(func() (string, error) {
+				return "", errExpected
 			})
 		})
 	})
 
-	t.Run("panic notnil result notnil error", func(t *testing.T) {
-		cmd := new(exec.Cmd)
+	t.Run("panic because error 2", func(t *testing.T) {
 		errExpected := errors.New("something went wrong")
 
 		assert.PanicsWithValue(t, errExpected, func() {
-			Must(func() (interface{}, error) {
-				return cmd, errExpected // Не делайте так в реальной жизни!
+			Must(func() (string, error) {
+				return "garbage", errExpected // Не делайте так в реальной жизни!
 			})
 		})
 	})
@@ -54,7 +62,7 @@ func TestMust(t *testing.T) {
 		expectedMessage := "something went really wrong"
 
 		assert.PanicsWithValue(t, expectedMessage, func() {
-			Must(func() (interface{}, error) {
+			Must(func() (any, error) {
 				panic(expectedMessage)
 			})
 		})
