@@ -1,7 +1,7 @@
 package panichelpers
 
 type Logger interface {
-	Logf(format string, args ...interface{})
+	Logf(format string, args ...any)
 }
 
 func LogPanicWithTrace(l Logger) {
