@@ -13,98 +13,98 @@ import (
 func TestRecover(t *testing.T) {
 	cases := []struct {
 		name   string
-		newArg func() interface{}
+		newArg func() any
 	}{
 		{
 			name:   "bool",
-			newArg: func() interface{} { return false },
+			newArg: func() any { return false },
 		},
 		{
 			name:   "int",
-			newArg: func() interface{} { return 42 },
+			newArg: func() any { return 42 },
 		},
 		{
 			name:   "int8",
-			newArg: func() interface{} { return int8(42) },
+			newArg: func() any { return int8(42) },
 		},
 		{
 			name:   "int16",
-			newArg: func() interface{} { return int16(42) },
+			newArg: func() any { return int16(42) },
 		},
 		{
 			name:   "int32",
-			newArg: func() interface{} { return int32(42) },
+			newArg: func() any { return int32(42) },
 		},
 		{
 			name:   "int64",
-			newArg: func() interface{} { return int64(42) },
+			newArg: func() any { return int64(42) },
 		},
 		{
 			name:   "uint",
-			newArg: func() interface{} { return uint(24) },
+			newArg: func() any { return uint(24) },
 		},
 		{
 			name:   "uint8",
-			newArg: func() interface{} { return uint8(24) },
+			newArg: func() any { return uint8(24) },
 		},
 		{
 			name:   "uint16",
-			newArg: func() interface{} { return uint16(24) },
+			newArg: func() any { return uint16(24) },
 		},
 		{
 			name:   "uint32",
-			newArg: func() interface{} { return uint32(24) },
+			newArg: func() any { return uint32(24) },
 		},
 		{
 			name:   "uint64",
-			newArg: func() interface{} { return uint64(24) },
+			newArg: func() any { return uint64(24) },
 		},
 		{
 			name:   "float32",
-			newArg: func() interface{} { return float32(42.) },
+			newArg: func() any { return float32(42.) },
 		},
 		{
 			name:   "float64",
-			newArg: func() interface{} { return 42. },
+			newArg: func() any { return 42. },
 		},
 		{
 			name:   "complex64",
-			newArg: func() interface{} { return complex(float32(1.), float32(10.)) },
+			newArg: func() any { return complex(float32(1.), float32(10.)) },
 		},
 		{
 			name:   "complex128",
-			newArg: func() interface{} { return complex(10., 100.) },
+			newArg: func() any { return complex(10., 100.) },
 		},
 		{
 			name:   "array",
-			newArg: func() interface{} { return [...]int{1, 2, 3, 4, 5} },
+			newArg: func() any { return [...]int{1, 2, 3, 4, 5} },
 		},
 		{
 			name:   "chan",
-			newArg: func() interface{} { return make(chan int, 100) },
+			newArg: func() any { return make(chan int, 100) },
 		},
 		{
 			name:   "interface",
-			newArg: func() interface{} { return error(new(runtime.TypeAssertionError)) },
+			newArg: func() any { return error(new(runtime.TypeAssertionError)) },
 		},
 		{
 			name: "ptr",
-			newArg: func() interface{} {
+			newArg: func() any {
 				i := 42
 				return &i
 			},
 		},
 		{
 			name:   "string",
-			newArg: func() interface{} { return "don't panic!" },
+			newArg: func() any { return "don't panic!" },
 		},
 		{
 			name:   "struct",
-			newArg: func() interface{} { return template.ExecError{} },
+			newArg: func() any { return template.ExecError{} },
 		},
 		{
 			name: "unsafe ptr",
-			newArg: func() interface{} {
+			newArg: func() any {
 				i := 42
 				return unsafe.Pointer(&i)
 			},
