@@ -2,7 +2,7 @@ package panichelpers
 
 // Recover вызывает fn, ловит панику и возвращает её аргумент.
 // Если паники не было, то функция возвращает <nil>.
-func Recover(fn func()) interface{} {
+func Recover(fn func()) any {
 	// Реализуй меня.
 	return nil
 }
