@@ -2,7 +2,7 @@ package fnhelpers
 
 // Must выполняет функцию fn и паникует, если та завершилась с ошибкой.
 // В обратном случае возвращает результат работы функции.
-func Must(fn func() (interface{}, error)) interface{} {
+func Must() {
+	// Исправь мою сигнатуру.
 	// Реализуй меня.
-	return nil
 }
